Move list header view and tableview into MyListViewController
