--- santaProject/MyList/MyListView.swift
@@ -1,45 +1,0 @@
-//
-//  MyListView.swift
-//  santaProject
-//
-//  Created by Juwon Kim on 2021/06/05.
-//
-
-import UIKit
-
-class MyListView: UIView {
-
-    let headerView = MyListHeaderView()
-
-    let tableView = MyListTableView(frame: .zero, style: .plain)
-
-    init() {
-        super.init(frame: .zero)
-        setupHeaderView()
-        setupTableView()
-
-        backgroundColor = UIColor(hex: "F0F0F0")
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    private func setupHeaderView() {
-        addSubview(headerView)
-        headerView.translatesAutoresizingMaskIntoConstraints = false
-        headerView.leadingAnchor.constraint(equalTo: leadingAnchor).isActive = true
-        headerView.trailingAnchor.constraint(equalTo: trailingAnchor).isActive = true
-        headerView.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor).isActive = true
-        headerView.heightAnchor.constraint(equalToConstant: 53).isActive = true
-    }
-
-    private func setupTableView() {
-        addSubview(tableView)
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        tableView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 20).isActive = true
-        tableView.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -20).isActive = true
-        tableView.topAnchor.constraint(equalTo: headerView.bottomAnchor, constant: 11).isActive = true
-        tableView.bottomAnchor.constraint(equalTo: safeAreaLayoutGuide.bottomAnchor).isActive = true
-    }
-}
